Extract the "smalls" stuff into its own base class.
...for potential reuse.
